requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
numpy
plotly
dash
PyYAML
alpaca-trade-api
scikit-learn
tensorflow
keras